Add delete confirmation to all the things.

+++ flutterapp/lib/components/taskitem.dart
@@ -1,3 +1,4 @@
+import 'package:docket/dialogs/confirmdelete.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -89,12 +90,17 @@ class TaskActions extends StatelessWidget {
     }
 
     Future<void> handleDelete() async {
-      try {
-        await tasksProvider.deleteTask(task);
-        messenger.showSnackBar(successSnackBar(context: context, text: 'Task Deleted'));
-      } catch (e) {
-        messenger.showSnackBar(errorSnackBar(context: context, text: 'Could not delete task'));
-      }
+      showConfirmDelete(
+        context: context, 
+        content: "Are you sure you want to delete this task?",
+        onConfirm: () async {
+        try {
+          await tasksProvider.deleteTask(task);
+          messenger.showSnackBar(successSnackBar(context: context, text: 'Task Deleted'));
+        } catch (e) {
+          messenger.showSnackBar(errorSnackBar(context: context, text: 'Could not delete task'));
+        }
+      });
     }
 
     Future<void> handleReschedule() async {
